+++ sudachi-cli/src/output.rs
@@ -15,7 +15,7 @@
  */
 
 use std::io::{BufWriter, Write};
-use sudachi::analysis::morpheme::Morpheme;
+use sudachi::analysis::morpheme::MorphemeView;
 use sudachi::dic::subset::InfoSubset;
 use sudachi::dic::DictionaryAccess;
 
@@ -106,7 +106,7 @@ impl<T: DictionaryAccess> SudachiOutput<T> for Simple {
 fn write_morpheme_basic<T, M>(writer: &mut Writer, morpheme: &M) -> SudachiResult<()>
 where
     T: DictionaryAccess,
-    M: Morpheme<Dictionary = T> + ?Sized,
+    M: MorphemeView<Dictionary = T> + ?Sized,
 {
     writer.write_all(morpheme.surface().as_bytes())?;
     writer.write_all(b"\t")?;
@@ -126,7 +126,7 @@ where
 fn write_morpheme_extended<T, M>(writer: &mut Writer, morpheme: &M) -> SudachiResult<()>
 where
     T: DictionaryAccess,
-    M: Morpheme<Dictionary = T> + ?Sized,
+    M: MorphemeView<Dictionary = T> + ?Sized,
 {
     write!(
         writer,